added console project
