+++ Handwriting/Handwriting.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30225.117
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Handwriting", "Handwriting\Handwriting.csproj", "{D327F6CE-4A46-4CA0-9AA3-B6A038232595}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HandwritingLibrary", "HandwritingLibrary\HandwritingLibrary.csproj", "{CB2143C5-B210-4122-A8C3-A5AF58C45203}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D327F6CE-4A46-4CA0-9AA3-B6A038232595}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D327F6CE-4A46-4CA0-9AA3-B6A038232595}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D327F6CE-4A46-4CA0-9AA3-B6A038232595}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D327F6CE-4A46-4CA0-9AA3-B6A038232595}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CB2143C5-B210-4122-A8C3-A5AF58C45203}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CB2143C5-B210-4122-A8C3-A5AF58C45203}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CB2143C5-B210-4122-A8C3-A5AF58C45203}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CB2143C5-B210-4122-A8C3-A5AF58C45203}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {79ECDE5D-DB7B-4DF1-A610-35AAE70C71C6}
+	EndGlobalSection
+EndGlobal
