+++ Handwriting/Handwriting/中.txt
@@ -1,0 +1,1 @@
+[[[39.011383056640625,150.79261779785156],[40.011383056640625,150.79261779785156],[41.011383056640625,164.79261779785156],[41.011383056640625,165.79261779785156],[41.011383056640625,169.79261779785156],[41.011383056640625,177.79261779785156],[41.011383056640625,178.79261779785156],[41.011383056640625,184.79261779785156],[41.011383056640625,196.79261779785156],[41.011383056640625,203.79261779785156],[41.011383056640625,206.79261779785156],[41.011383056640625,208.79261779785156],[41.011383056640625,210.79261779785156],[41.011383056640625,212.79261779785156],[41.011383056640625,215.79261779785156]],[[41.011383056640625,147.79261779785156],[41.011383056640625,145.79261779785156],[44.011383056640625,144.79261779785156],[47.011383056640625,144.79261779785156],[56.011383056640625,144.79261779785156],[58.011383056640625,144.79261779785156],[65.01138305664062,143.79261779785156],[69.01138305664062,142.79261779785156],[75.01138305664062,142.79261779785156],[83.01138305664062,142.79261779785156],[91.01138305664062,142.79261779785156],[97.01138305664062,142.79261779785156],[98.01138305664062,142.79261779785156],[100.01138305664062,142.79261779785156],[103.01138305664062,142.79261779785156],[104.01138305664062,142.79261779785156],[105.01138305664062,142.79261779785156],[107.01138305664062,142.79261779785156],[109.01138305664062,142.79261779785156],[113.01138305664062,142.79261779785156],[117.01138305664062,142.79261779785156],[122.01138305664062,142.79261779785156],[128.01138305664062,142.79261779785156],[132.01138305664062,142.79261779785156],[135.01138305664062,142.79261779785156],[136.01138305664062,142.79261779785156],[137.01138305664062,142.79261779785156],[138.01138305664062,144.79261779785156],[138.01138305664062,145.79261779785156],[138.01138305664062,146.79261779785156],[138.01138305664062,148.79261779785156],[138.01138305664062,148.79261779785156],[138.01138305664062,149.79261779785156],[138.01138305664062,151.79261779785156],[138.01138305664062,153.79261779785156],[138.01138305664062,155.79261779785156],[138.01138305664062,157.79261779785156],[138.01138305664062,163.79261779785156],[138.01138305664062,168.79261779785156],[137.01138305664062,174.79261779785156],[136.01138305664062,175.79261779785156],[136.01138305664062,177.79261779785156],[136.01138305664062,178.79261779785156],[136.01138305664062,181.79261779785156],[136.01138305664062,186.79261779785156],[135.01138305664062,189.79261779785156],[135.01138305664062,197.79261779785156],[135.01138305664062,199.79261779785156],[135.01138305664062,205.79261779785156]],[[43.011383056640625,205.79261779785156],[45.011383056640625,205.79261779785156],[49.011383056640625,205.79261779785156],[53.011383056640625,204.79261779785156],[60.011383056640625,204.79261779785156],[63.011383056640625,204.79261779785156],[78.01138305664062,204.79261779785156],[85.01138305664062,204.79261779785156],[95.01138305664062,203.79261779785156],[99.01138305664062,202.79261779785156],[108.01138305664062,201.79261779785156],[110.01138305664062,201.79261779785156],[112.01138305664062,200.79261779785156],[115.01138305664062,200.79261779785156],[117.01138305664062,200.79261779785156],[118.01138305664062,200.79261779785156],[122.01138305664062,200.79261779785156],[127.01138305664062,200.79261779785156],[128.01138305664062,200.79261779785156],[128.01138305664062,199.79261779785156]],[[91.01138305664062,115.79261779785156],[90.01138305664062,115.79261779785156],[90.01138305664062,117.79261779785156],[90.01138305664062,118.79261779785156],[90.01138305664062,122.79261779785156],[90.01138305664062,126.79261779785156],[90.01138305664062,131.79261779785156],[90.01138305664062,137.79261779785156],[90.01138305664062,145.79261779785156],[89.01138305664062,154.79261779785156],[88.01138305664062,161.79261779785156],[87.01138305664062,171.79261779785156],[86.01138305664062,190.79261779785156],[85.01138305664062,199.79261779785156],[83.01138305664062,212.79261779785156],[83.01138305664062,227.79261779785156],[82.01138305664062,232.79261779785156],[82.01138305664062,238.79261779785156],[82.01138305664062,247.79261779785156],[82.01138305664062,251.79261779785156],[82.01138305664062,255.79261779785156],[82.01138305664062,258.79261779785156],[82.01138305664062,259.79261779785156],[82.01138305664062,260.79261779785156],[82.01138305664062,261.79261779785156],[82.01138305664062,262.79261779785156]]]
