+++ Handwriting/Handwriting/也.txt
@@ -1,0 +1,1 @@
+[[[77.01138305664062,163.79261779785156],[78.01138305664062,164.79261779785156],[111.01138305664062,158.79261779785156],[113.01138305664062,158.79261779785156],[119.01138305664062,157.79261779785156],[124.01138305664062,156.79261779785156],[128.01138305664062,154.79261779785156],[131.01138305664062,154.79261779785156],[133.01138305664062,153.79261779785156],[138.01138305664062,152.79261779785156],[139.01138305664062,151.79261779785156],[142.01138305664062,151.79261779785156],[150.01138305664062,149.79261779785156],[153.01138305664062,148.79261779785156],[155.01138305664062,148.79261779785156],[163.01138305664062,146.79261779785156],[166.01138305664062,145.79261779785156],[169.01138305664062,144.79261779785156],[171.01138305664062,143.79261779785156],[174.01138305664062,143.79261779785156],[178.01138305664062,141.79261779785156],[179.01138305664062,140.79261779785156],[182.01138305664062,140.79261779785156],[185.01138305664062,140.79261779785156],[187.01138305664062,140.79261779785156],[187.01138305664062,140.79261779785156],[188.01138305664062,140.79261779785156],[189.01138305664062,139.79261779785156],[190.01138305664062,139.79261779785156],[192.01138305664062,138.79261779785156],[194.01138305664062,138.79261779785156],[195.01138305664062,138.79261779785156],[195.01138305664062,139.79261779785156],[195.01138305664062,140.79261779785156],[195.01138305664062,142.79261779785156],[195.01138305664062,146.79261779785156],[195.01138305664062,147.79261779785156],[195.01138305664062,148.79261779785156],[195.01138305664062,153.79261779785156],[195.01138305664062,155.79261779785156],[195.01138305664062,158.79261779785156],[195.01138305664062,160.79261779785156],[195.01138305664062,161.79261779785156],[195.01138305664062,163.79261779785156],[195.01138305664062,164.79261779785156],[195.01138305664062,166.79261779785156],[195.01138305664062,170.79261779785156],[195.01138305664062,175.79261779785156],[195.01138305664062,176.79261779785156],[195.01138305664062,178.79261779785156],[195.01138305664062,180.79261779785156],[195.01138305664062,181.79261779785156],[195.01138305664062,182.79261779785156],[195.01138305664062,186.79261779785156],[195.01138305664062,188.79261779785156],[195.01138305664062,189.79261779785156],[195.01138305664062,191.79261779785156],[195.01138305664062,192.79261779785156],[194.01138305664062,193.79261779785156],[194.01138305664062,194.79261779785156],[194.01138305664062,195.79261779785156],[194.01138305664062,196.79261779785156],[194.01138305664062,197.79261779785156],[194.01138305664062,198.79261779785156],[194.01138305664062,201.79261779785156],[194.01138305664062,202.79261779785156],[194.01138305664062,204.79261779785156],[194.01138305664062,205.79261779785156],[194.01138305664062,207.79261779785156],[192.01138305664062,208.79261779785156],[192.01138305664062,208.79261779785156],[192.01138305664062,209.79261779785156],[192.01138305664062,210.79261779785156],[191.01138305664062,211.79261779785156],[190.01138305664062,211.79261779785156],[189.01138305664062,211.79261779785156],[188.01138305664062,211.79261779785156],[186.01138305664062,211.79261779785156],[184.01138305664062,211.79261779785156],[183.01138305664062,209.79261779785156]],[[154.01138305664062,119.79261779785156],[153.01138305664062,119.79261779785156],[153.01138305664062,121.79261779785156],[153.01138305664062,124.79261779785156],[153.01138305664062,135.79261779785156],[153.01138305664062,139.79261779785156],[153.01138305664062,149.79261779785156],[153.01138305664062,150.79261779785156],[153.01138305664062,153.79261779785156],[153.01138305664062,156.79261779785156],[153.01138305664062,159.79261779785156],[153.01138305664062,163.79261779785156],[153.01138305664062,164.79261779785156],[153.01138305664062,168.79261779785156],[153.01138305664062,170.79261779785156],[153.01138305664062,177.79261779785156],[153.01138305664062,185.79261779785156],[153.01138305664062,191.79261779785156],[153.01138305664062,193.79261779785156],[154.01138305664062,197.79261779785156],[154.01138305664062,198.79261779785156],[154.01138305664062,201.79261779785156],[154.01138305664062,204.79261779785156],[155.01138305664062,205.79261779785156],[156.01138305664062,208.79261779785156],[157.01138305664062,210.79261779785156],[157.01138305664062,214.79261779785156],[157.01138305664062,215.79261779785156]],[[115.01138305664062,112.79261779785156],[115.01138305664062,113.79261779785156],[115.01138305664062,115.79261779785156],[115.01138305664062,117.79261779785156],[115.01138305664062,119.79261779785156],[115.01138305664062,132.79261779785156],[115.01138305664062,134.79261779785156],[115.01138305664062,139.79261779785156],[115.01138305664062,147.79261779785156],[117.01138305664062,153.79261779785156],[118.01138305664062,158.79261779785156],[119.01138305664062,163.79261779785156],[119.01138305664062,168.79261779785156],[120.01138305664062,175.79261779785156],[122.01138305664062,182.79261779785156],[123.01138305664062,187.79261779785156],[124.01138305664062,195.79261779785156],[125.01138305664062,198.79261779785156],[128.01138305664062,212.79261779785156],[128.01138305664062,219.79261779785156],[128.01138305664062,228.79261779785156],[129.01138305664062,234.79261779785156],[129.01138305664062,246.79261779785156],[129.01138305664062,248.79261779785156],[129.01138305664062,255.79261779785156],[129.01138305664062,262.79261779785156],[130.01138305664062,269.79261779785156],[130.01138305664062,271.79261779785156],[130.01138305664062,277.79261779785156],[132.01138305664062,285.79261779785156],[132.01138305664062,286.79261779785156],[134.01138305664062,289.79261779785156],[134.01138305664062,290.79261779785156],[134.01138305664062,291.79261779785156],[135.01138305664062,292.79261779785156],[138.01138305664062,292.79261779785156],[140.01138305664062,292.79261779785156],[147.01138305664062,290.79261779785156],[152.01138305664062,288.79261779785156],[155.01138305664062,288.79261779785156],[160.01138305664062,286.79261779785156],[169.01138305664062,285.79261779785156],[179.01138305664062,282.79261779785156],[188.01138305664062,279.79261779785156],[191.01138305664062,278.79261779785156],[197.01138305664062,277.79261779785156],[202.01138305664062,275.79261779785156],[206.01138305664062,273.79261779785156],[207.01138305664062,271.79261779785156],[209.01138305664062,270.79261779785156],[210.01138305664062,269.79261779785156],[211.01138305664062,268.79261779785156],[213.01138305664062,267.79261779785156],[214.01138305664062,267.79261779785156],[216.01138305664062,265.79261779785156],[217.01138305664062,263.79261779785156],[218.01138305664062,262.79261779785156],[218.01138305664062,260.79261779785156],[219.01138305664062,258.79261779785156],[220.01138305664062,258.79261779785156],[220.01138305664062,255.79261779785156],[220.01138305664062,254.79261779785156],[220.01138305664062,252.79261779785156],[220.01138305664062,250.79261779785156],[220.01138305664062,249.79261779785156],[220.01138305664062,248.79261779785156],[220.01138305664062,247.79261779785156],[220.01138305664062,245.79261779785156]]]
